+++ api/lib/application/organization-learner-user-associations/index.js
@@ -9,37 +9,6 @@ exports.register = async function (server) {
   server.route([
     {
       method: 'POST',
-      path: '/api/schooling-registration-user-associations',
-      config: {
-        handler: organizationLearnerUserAssociationController.reconcileOrganizationLearnerManually,
-        validate: {
-          options: {
-            allowUnknown: false,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'first-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                'last-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                birthdate: Joi.date().format('YYYY-MM-DD').required(),
-                'campaign-code': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-              },
-              type: 'schooling-registration-user-associations',
-            },
-          }),
-          failAction: (request, h) => {
-            return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
-          },
-        },
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Elle associe des données saisies par l’utilisateur à l’inscription de l’élève dans cette organisation',
-        ],
-        tags: ['api', 'organizationLearnerUserAssociation'],
-      },
-    },
-    {
-      method: 'POST',
       path: '/api/schooling-registration-user-associations/student',
       config: {
         handler: organizationLearnerUserAssociationController.reconcileSupOrganizationLearner,
