+++ api/lib/domain/usecases/index.js
@@ -363,8 +363,8 @@ module.exports = injectDependencies(
     neutralizeChallenge: require('./neutralize-challenge'),
     publishSession: require('./publish-session'),
     publishSessionsInBatch: require('./publish-sessions-in-batch'),
+    reconcileScoOrganizationLearnerManually: require('./reconcile-sco-organization-learner-manually'),
     reconcileSupOrganizationLearner: require('./reconcile-sup-organization-learner'),
-    reconcileOrganizationLearner: require('./reconcile-organization-learner'),
     reconcileUserToOrganization: require('./reconcile-user-to-organization'),
     rememberUserHasSeenAssessmentInstructions: require('./remember-user-has-seen-assessment-instructions'),
     rememberUserHasSeenChallengeTooltip: require('./remember-user-has-seen-challenge-tooltip'),
