+++ api/lib/infrastructure/serializers/jsonapi/sco-organization-learner-serializer.js
@@ -1,0 +1,9 @@
+const { Serializer } = require('jsonapi-serializer');
+
+module.exports = {
+  serializeIdentity(scoOrganizationLearner) {
+    return new Serializer('sco-organization-learner', {
+      attributes: ['lastName', 'firstName', 'birthdate'],
+    }).serialize(scoOrganizationLearner);
+  },
+};
