+++ api/tests/integration/application/sco-organization-learners/index_test.js
@@ -1,0 +1,385 @@
+const { expect, sinon, HttpTestServer } = require('../../../test-helper');
+
+const scoOrganizationLearnerController = require('../../../../lib/application/sco-organization-learners/sco-organization-learner-controller');
+const moduleUnderTest = require('../../../../lib/application/sco-organization-learners');
+
+describe('Integration | Application | Route | sco-organization-learners', function () {
+  let httpTestServer;
+
+  beforeEach(async function () {
+    sinon
+      .stub(scoOrganizationLearnerController, 'reconcileScoOrganizationLearnerManually')
+      .callsFake((request, h) => h.response('ok').code(204));
+
+    httpTestServer = new HttpTestServer();
+    await httpTestServer.register(moduleUnderTest);
+  });
+
+  describe('POST /api/schooling-registration-user-associations', function () {
+    const method = 'POST';
+    const url = '/api/schooling-registration-user-associations';
+
+    context('User association with firstName, lastName, birthdate and campaignCode', function () {
+      it('should succeed', async function () {
+        // given
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
+
+      it('should succeed when there is a space', async function () {
+        // given
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert ',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+        expect(response.request.payload.data.attributes['first-name']).to.equal('Robert ');
+      });
+
+      it('should return an error when there is no payload', async function () {
+        // when
+        const response = await httpTestServer.request(method, url);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid first name attribute in the payload', async function () {
+        // given
+        const INVALID_FIRSTNAME = ' ';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': INVALID_FIRSTNAME,
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid last name attribute in the payload', async function () {
+        // given
+        const INVALID_LASTNAME = '';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': INVALID_LASTNAME,
+              birthdate: '2012-12-12',
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid a birthdate attribute (with space) in the payload', async function () {
+        // given
+        const INVALID_BIRTHDATE = '2012- 12-12';
+
+        // when
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: INVALID_BIRTHDATE,
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid birthdate attribute (with extra zeros) in the payload', async function () {
+        // given
+        const INVALID_BIRTHDATE = '2012-012-12';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: INVALID_BIRTHDATE,
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid birthdate attribute (not a proper date) in the payload', async function () {
+        // given
+        const INVALID_BIRTHDATE = '1999-99-99';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: INVALID_BIRTHDATE,
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid campaign code attribute in the payload', async function () {
+        // given
+        const INVALID_CAMPAIGNCODE = '';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              'campaign-code': INVALID_CAMPAIGNCODE,
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+    });
+  });
+
+  describe('POST /api/sco-organization-learners/association', function () {
+    const method = 'POST';
+    const url = '/api/sco-organization-learners/association';
+
+    context('User association with firstName, lastName, birthdate and campaignCode', function () {
+      it('should succeed', async function () {
+        // given
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
+
+      it('should succeed when there is a space', async function () {
+        // given
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert ',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+        expect(response.request.payload.data.attributes['first-name']).to.equal('Robert ');
+      });
+
+      it('should return an error when there is no payload', async function () {
+        // when
+        const response = await httpTestServer.request(method, url);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid first name attribute in the payload', async function () {
+        // given
+        const INVALID_FIRSTNAME = ' ';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': INVALID_FIRSTNAME,
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid last name attribute in the payload', async function () {
+        // given
+        const INVALID_LASTNAME = '';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': INVALID_LASTNAME,
+              birthdate: '2012-12-12',
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid a birthdate attribute (with space) in the payload', async function () {
+        // given
+        const INVALID_BIRTHDATE = '2012- 12-12';
+
+        // when
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: INVALID_BIRTHDATE,
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid birthdate attribute (with extra zeros) in the payload', async function () {
+        // given
+        const INVALID_BIRTHDATE = '2012-012-12';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: INVALID_BIRTHDATE,
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid birthdate attribute (not a proper date) in the payload', async function () {
+        // given
+        const INVALID_BIRTHDATE = '1999-99-99';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: INVALID_BIRTHDATE,
+              'campaign-code': 'RESTRICTD',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+
+      it('should return an error when there is an invalid campaign code attribute in the payload', async function () {
+        // given
+        const INVALID_CAMPAIGNCODE = '';
+        const payload = {
+          data: {
+            attributes: {
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              'campaign-code': INVALID_CAMPAIGNCODE,
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(422);
+      });
+    });
+  });
+});
